fixed typo in progress.cpp: int->size_t
From: Alexander Petry <alexander.petry@itwm.fraunhofer.de>

git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@7006 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/sdpa/trunk/plugin/progress.cpp
@@ -84,7 +84,7 @@ private:
   std::string m_prefix;
 };
 
-int set_progress(const char *name, int value)
+int set_progress(const char *name, size_t value)
 {
   return global_progress->set(name, value);
 }
